// src/model/user.ts
export type IUserInfo = {
  anchorCoin: number;
  anchorExp: number;
  anchorLevel: number;
  anchorLiveNotice: number;
  auth: boolean;
  autoUpdownBalance: number;
  avatar: string;
  badgeList: number[];
  birthday: string;
  chatHide: number;
  city: string;
  fans: number;
  follows: number;
  gameQuota: number;
  goldCoin: number;
  imToken: string;
  isBlackChat: boolean;
  isFollow: boolean;
  isReject: boolean;
  isRename: boolean;
  liveId: string;
  liveStatus: number;
  manage: number;
  nickname: string;
  phone: string;
  rankHide: number;
  receiveCoin: number;
  roomHide: number;
  sendCoin: number;
  sex: number;
  signature: string;
  uid: number;
  userExp: number;
  userLevel: number;
  vipUid: number;
};

export const initUser: IUserInfo = {
  anchorCoin: 0,
  anchorExp: 0,
  anchorLevel: 0,
  anchorLiveNotice: 0,
  auth: false,
  autoUpdownBalance: 0,
  avatar: "",
  badgeList: [],
  birthday: "",
  chatHide: 0,
  city: "",
  fans: 0,
  follows: 0,
  gameQuota: 0,
  goldCoin: 0,
  imToken: "",
  isBlackChat: false,
  isFollow: false,
  isReject: false,
  isRename: false,
  liveId: "",
  liveStatus: 0,
  manage: 0,
  nickname: "",
  phone: "",
  rankHide: 0,
  receiveCoin: 0,
  roomHide: 0,
  sendCoin: 0,
  sex: 0,
  signature: "",
  uid: 0,
  userExp: 0,
  userLevel: 0,
  vipUid: 0,
};